introduced hints the client can give to the core
The client can now tell the core that trap instructions
are likely to return and continue at the next instruction,
so the core will translate the code after the trap eagerly.

OpenBSD on M88K returns to pc+4 or pc+8 after a syscall,
depending on whether there was an error. The client can hint
this as well, which makes the core treat both pc+4 and pc+8
return addresses.