moved entry caching into the core
Tagging is greedy again. We tag as much as possible, and take hints from
the cache file. Whenever we encounter new code, we append it to the cache
on disk.
This strategy causes significant latency when starting a program, but
on subsequent runs, this latency will only be at the start, and all
translation sequences while running should go away...