libnix: Fix setting/getting termios baud rate on musl
Use cfget*/cfset* on all systems, instead of non-standardized c_ispeed
and c_ospeed members of struct termios.

In some implementations these fields are private or don't even exist,
but the functions to get/set them should be available on all systems,
or at least could be easily tested for and conditionally provided when
missing, which is a more robust and portable way to handle it.

Closes: https://github.com/libcpu/libcpu/issues/10

Signed-off-by: Ismael Luceno <ismael@iodev.co.uk>